feat(*): add support for github action cache (#5)
Add a new option to enable the usage of GitHub Action Cache to cache the
Kong test image and avoid rebuilding it at each run.

Note that it takes a bit longer on the first run as the layer needs to
be exported. It then divides by 2 the execution time of the action for
subsequent executions.